# docs/openapi/turbospark.openapi.yaml
openapi: 3.1.0
info:
  title: TurboSpark Server API
  version: 0.2.0
  description: |
    Local inference HTTP server (turbospark-server / turbospark serve).

    Provides standardized endpoints for Apple Silicon inference:
    - OpenAI-compatible chat completions, prompt completions, responses, embeddings, and image generation
    - Anthropic-compatible messages and token counting
    - Ollama-compatible chat, generate, tags, version, show, and embeddings
    - Audio capabilities: speech-to-text (Whisper, Qwen3-ASR), text-to-speech (Kokoro), music generation (MiniMax Music 0.3), and realtime WebSocket transcription
    - System health probes and Prometheus metrics

    Supported transport features:
    - SSE streaming for OpenAI and Anthropic routes
    - NDJSON streaming for Ollama and audio transcription streams
    - API key authentication via x-api-key or Authorization: Bearer
servers:
  - url: http://127.0.0.1:8080
    description: Default local server
security:
  - bearer: []
  - apiKey: []
paths:
...
  /v1/chat/completions:
    post:
      summary: Create chat completion
      description: |
        OpenAI-compatible chat completion endpoint. Supports text generation,
        reasoning channels, structured tool calling, and multimodal image input.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: |
            Successful completion. Returns JSON when stream=false, or an SSE stream
            of ChatCompletionChunk objects when stream=true.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Busy' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/completions:
    post:
      summary: Raw prompt completion
      description: |
        OpenAI legacy raw-prompt completion. Encodes the prompt directly without
        applying a chat template.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletionRequest'
      responses:
        '200':
          description: Completion response or SSE stream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletionResponse'
            text/event-stream:
              schema:
                type: string
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Busy' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/responses:
    post:
      summary: OpenAI Responses API
      description: Item-shaped request and response endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponsesRequest'
      responses:
        '200':
          description: Response or SSE stream of ResponsesStreamEvent
          content:
            application/json:
              schema:
                type: object
            text/event-stream:
              schema:
                type: string
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Busy' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/messages:
    post:
      summary: Anthropic messages completion
      description: |
        Anthropic-compatible messages endpoint. Automatically translated to/from
        the core generation engine.
      parameters:
        - name: anthropic-version
          in: header
          schema:
            type: string
          description: Anthropic API version tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnthropicMessageRequest'
      responses:
        '200':
          description: Message response or SSE stream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnthropicMessageResponse'
            text/event-stream:
              schema:
                type: string
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Busy' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/messages/count_tokens:
    post:
      summary: Anthropic token count probe
      description: Counts prompt tokens without running generation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnthropicMessageRequest'
      responses:
        '200':
          description: Input token count
          content:
            application/json:
              schema:
                type: object
                properties:
                  input_tokens:
                    type: integer
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /v1/models/{model}:
    get:
      summary: Retrieve model details
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '404': { $ref: '#/components/responses/Error' }
  /v1/embeddings:
    post:
      summary: Create embeddings
      description: |
        Generates vector embeddings for input text strings. Supported encoding
        formats include float and base64.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '200':
          description: Embedding vectors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Busy' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/images/generations:
    post:
      summary: Generate image from prompt
      description: |
        Prompt-to-PNG generation using attached diffusion models (e.g. Z-Image,
        Qwen-Image). Produces base64-encoded PNG data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageGenerateRequest'
      responses:
        '200':
          description: Generated images
          headers:
            x-turbospark-seed:
              schema:
                type: integer
              description: First resolved generation seed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Busy' }
        '503': { $ref: '#/components/responses/Error' }
...
  /api/show:
    post:
      summary: Ollama show model details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Model metadata
          content:
            application/json:
              schema:
                type: object
        '404': { $ref: '#/components/responses/Error' }
  /api/chat:
    post:
      summary: Ollama chat completion
      description: NDJSON-streamed chat completion in Ollama format.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OllamaChatRequest'
      responses:
        '200':
          description: NDJSON stream or JSON object
          content:
            application/x-ndjson:
              schema:
                type: string
            application/json:
              schema:
                type: object
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/generate:
    post:
      summary: Ollama raw generate
      description: NDJSON-streamed prompt generation in Ollama format.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OllamaGenerateRequest'
      responses:
        '200':
          description: NDJSON stream or JSON object
          content:
            application/x-ndjson:
              schema:
                type: string
            application/json:
              schema:
                type: object
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/embeddings:
    post:
      summary: Ollama legacy embeddings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, prompt]
              properties:
                model: { type: string }
                prompt: { type: string }
      responses:
        '200':
          description: Embedding vector
          content:
            application/json:
              schema:
                type: object
                properties:
                  embedding:
                    type: array
                    items: { type: number }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/embed:
    post:
      summary: Ollama batch embeddings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, input]
              properties:
                model: { type: string }
                input:
                  oneOf:
                    - type: string
                    - type: array
                      items: { type: string }
      responses:
        '200':
          description: Batch embeddings
          content:
            application/json:
              schema:
                type: object
                properties:
                  embeddings:
                    type: array
                    items:
                      type: array
                      items: { type: number }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/audio/transcriptions:
    post:
      summary: Transcribe audio to text
      description: |
        Send multipart/form-data or raw audio/wav body. Supported models include
        Whisper and Qwen3-ASR.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TranscriptionForm'
          audio/wav:
            schema:
              type: string
              format: binary
      parameters:
        - name: model
          in: query
          schema: { type: string }
          description: Required for raw binary bodies
        - name: language
          in: query
          schema: { type: string }
        - name: response_format
          in: query
          schema: { $ref: '#/components/schemas/TranscriptionFormat' }
        - name: stream
          in: query
          schema: { type: boolean }
        - name: async
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Transcription result
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Transcription'
                  - $ref: '#/components/schemas/VerboseTranscription'
            text/plain: { schema: { type: string } }
            text/vtt: { schema: { type: string } }
            application/x-ndjson: { schema: { type: string } }
        '202':
          description: Accepted as asynchronous job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobAccepted' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
        '415': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Busy' }
        '500': { $ref: '#/components/responses/Error' }
...
  /v1/audio/transcriptions/realtime:
    get:
      summary: Utterance-level realtime transcription (WebSocket)
      description: |
        WebSocket upgrade endpoint for realtime 16 kHz mono PCM transcription.
        Send binary audio chunks followed by {"type":"commit"}.
      parameters:
        - name: model
          in: query
          required: true
          schema: { type: string }
        - name: encoding
          in: query
          schema:
            type: string
            enum: [s16le, f32le]
            default: s16le
        - name: language
          in: query
          schema: { type: string }
      responses:
        '101': { description: Switching Protocols }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Busy' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    ChatCompletionRequest:
      type: object
      required: [messages]
      properties:
        model: { type: string }
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        max_tokens: { type: integer }
        temperature: { type: number }
        top_p: { type: number }
        stream: { type: boolean, default: false }
        stop:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        tools:
          type: array
          items:
            type: object
        tool_choice:
          oneOf:
            - type: string
            - type: object
...
    TranscriptionFormat:
      type: string
      enum: [json, verbose_json, text, srt, vtt]
      default: json
...